+++ gr-blocks/lib/message_strobe_random_impl.cc
@@ -71,15 +71,15 @@ namespace gr {
                       boost::bind(&message_strobe_random_impl::set_msg, this, _1));
     }
 
-    float message_strobe_random_impl::next_delay(){
+    long message_strobe_random_impl::next_delay(){
         switch(d_dist){
             case STROBE_POISSON:
                 //return d_variate_poisson->operator()();
-                return d_variate_poisson->operator()();
+                return static_cast<long>(d_variate_poisson->operator()());
             case STROBE_GAUSSIAN:
-                return d_variate_normal->operator()();
+                return static_cast<long>(d_variate_normal->operator()());
             case STROBE_UNIFORM:
-                return d_variate_uniform->operator()();
+                return static_cast<long>(d_variate_uniform->operator()());
             default:
                 throw std::runtime_error("message_strobe_random_impl::d_distribution is very unhappy with you");
         }
@@ -110,7 +110,7 @@ namespace gr {
     void message_strobe_random_impl::run()
     {
       while(!d_finished) {
-        boost::this_thread::sleep(boost::posix_time::milliseconds(static_cast<long>(std::max(0.0f,next_delay()))));
+        boost::this_thread::sleep(boost::posix_time::milliseconds(std::max(0L, next_delay())));
         if(d_finished) {
           return;
         }
